Add local filesystem server
