--- examples/mcp_servers/local_filesystem_server/.env.example
@@ -1,1 +1,0 @@
-MY_SECRET_KEY="Your tools can have secrets injected at runtime!"
